create keybindings.go
